Добавлены предметы, переход на следующий уровень, level loading improved. Баги: с предметами творится полная хрень; неправильная обработка движения бэкграунда в levelLoop().

+++ object.h
@@ -12,7 +12,7 @@ typedef struct
 {
     char    isEnabled;
 
-    float   posX, posY,
+    int     posX, posY,
             destX, destY,
             reservedX, reservedY,
             speed;
